% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AD}
\alias{AD}
\title{Anterodorsal (AD) nucleus of the anterior thalamic nuclei}
\format{
An object of class \code{Seurat} with 297 rows and 724 columns.
}
\usage{
AD
}
\description{
Contains subset of Xenium data, including clusters spatially associated
  with the anterodorsal nucleus of the ATN. FOVs included: fov and X1fov,
}
\details{
Format: Seurat Object, with UMAP_1 embeddings computed and stored within.

Created in-house from Cembrowski Lab Xenium run to serve as example data.

 To load: data(AD) note: Lazy loading.
}
\keyword{datasets}
