% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Plot_geneExpressionvsMidline.R
\name{getExpressionvsMidline}
\alias{getExpressionvsMidline}
\title{Generate Gene Expression values for cells and their distance away from Spatial Midline}
\usage{
getExpressionvsMidline(object, FOV = "fov", genes, degs)
}
\arguments{
\item{object}{a Xenium object}

\item{FOV}{FOV to extract coordinates from}

\item{genes}{Character vector of gene names to fetch expression data for}

\item{degs}{User-defined angle, defined in degrees, that intersects centrepoint in midline computation}
}
\value{
Dataframe containing Cell IDs, coordinates (X,Y), gene expression counts for specififed gene(s),
   and computed distance away from spatial midline of each cell, in specified FOV.
}
\description{
Generates gene expression data away from the computed spatial midline. Supports one FOV at a time.
}
\examples{
geneExpression_X1fov <- getExpressionvsMidline(AD, FOV = "X1fov", genes = c("Epha1", "Cntnap4", 'Pcp4', 'Slc17a7'), degs = 45)
geneExpression_fov <- getExpressionvsMidline(AD, FOV = "fov", genes = c("Epha1", "Cntnap4", 'Pcp4', 'Slc17a7'), degs = 50)
}
\author{
Margarita Kapustina
}
