% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP1_line_and_histogram.R
\name{getUMAP1_MidlineData}
\alias{getUMAP1_MidlineData}
\title{Fetch UMAP_1 embeddings values for cells and their distance away from Spatial Midline}
\usage{
getUMAP1_MidlineData(object, FOV, degs = 45, binNumber = 15, binwidth = 40)
}
\arguments{
\item{object}{a Xenium object}

\item{FOV}{FOV to extract coordinates and UMAP_1 embeddings values from}

\item{degs}{User-defined angle, defined in degrees, that intersects centrepoint in midline computation}

\item{binNumber}{Number of bins to use for histogram (bars). Suggested binNumber = total # cells in pooled data/binwidth}

\item{binwidth}{Width of bins for gene expression averaging (lines) Suggested binwidth = 40 for 40micron bins}
}
\value{
Dataframe containing Cell IDs, coordinates (X,Y), UMAP_1 embedding values,
   and computed distance away from spatial midline of each cell, in specified FOV.
}
\description{
Fetches UMAP_1 embeddinga values for cells, and their distance away from the computed spatial midline. Supports one FOV at a time.
}
\examples{
UMAP1_midline_data_fov <- getUMAP1_MidlineData(AD, FOV= c('fov'), degs = 65)
UMAP1_midline_data_X1fov <- getUMAP1_MidlineData(AD, FOV= c('X1fov'), degs = 45)
}
\author{
Margarita Kapustina
}
