% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightCells_highlightCluster.R
\name{highlightCells}
\alias{highlightCells}
\title{Highlight Cells}
\usage{
highlightCells(
  highlight_obj,
  within_obj,
  size = 3,
  alpha_value = 0.5,
  FOV = c("fov", "X1fov"),
  color_palette = "glasbey",
  save_plot = FALSE
)
}
\arguments{
\item{highlight_obj}{Xenium object that you want to highlight cells from, labelled: "these cells"}

\item{within_obj}{Xenium object that you want show highlighted cells within, labelled: "all else" (ie cells plotted but not highlighted)}

\item{size}{Size of the highlighted cells in the plot}

\item{alpha_value}{Alpha (transparency) value of the highlighted cells}

\item{FOV}{Character vector specifying the fields of view (FOVs) to use for plotting}

\item{color_palette}{Color palette for the plot}

\item{save_plot}{Option to save plot as .pdf in working directory (TRUE, FALSE)}
}
\value{
An ImageDimPlot with highlighted cells.
}
\description{
Highlight cells from a Xenium object of interest, within another Xenium object, within specified FOV(s).
   Cells from object of interest labelled as "these cells", other cells labelled as "all else". Suggested to use when subsetting Xenium objects.
}
\examples{
highlightCells(highlight_obj = atn, within_obj = ExcitatoryNeurons, FOV = c('X1fov'))

}
\author{
Margarita Kapustina
}
