% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightCells_highlightCluster.R
\name{highlightCluster}
\alias{highlightCluster}
\title{Highlight Cluster}
\usage{
highlightCluster(
  object,
  cluster_id,
  size = 3,
  alpha_value = 0.5,
  FOV = c("fov", "X1fov"),
  color_palette = "glasbey",
  save_plot = FALSE
)
}
\arguments{
\item{object}{a Xenium object}

\item{cluster_id}{Identity of cluster to highlight}

\item{size}{Size of the highlighted cells in the plot}

\item{alpha_value}{Alpha (transparency) value of the highlighted cells}

\item{FOV}{Character vector specifying the fields of view (FOVs) to use for plotting}

\item{color_palette}{Color palette for the plot}

\item{save_plot}{Option to save plot as .pdf in working directory (TRUE, FALSE)}
}
\value{
An ImageDimPlot with highlighted cells.
}
\description{
Highlight cluster(s) of choice in a Xenium object, within specified FOV(s). Cluster of interest labelled "this cluster"
     and all other clusters labelled "all else". Supported for clusters within the same object.
}
\examples{
highlightCluster(obj, cluster_id = 1, FOV = c('X1fov', 'X2fov'))

}
\author{
Margarita Kapustina
}
