% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectToCoords.R
\name{object_FOV_to_coordinates}
\alias{object_FOV_to_coordinates}
\title{Get cell coordinates from a Xenium object}
\usage{
object_FOV_to_coordinates(
  object = object,
  thisFOV = "X9fov",
  threshold_y = 2000,
  angle_adjust = FALSE,
  theta_deg = 0,
  flip_x_coordinates = FALSE
)
}
\arguments{
\item{object}{Xenium object}

\item{thisFOV}{Name of your FOV to extract coordinates from}

\item{threshold_y}{Boundary that you define to compute cells that are above and below this boundary}

\item{angle_adjust}{Option to adjust the angle of your cell coordinates (TRUE, FALSE)}

\item{flip_x_coordinates}{Option to flip your slice in the horizontal plane (TRUE, FALSE)}

\item{thetda_deg}{Specify the degrees you wish to rotate your slice by}
}
\value{
A dataframe with cell ID, coordinates, cluster ID, and a plot with UMAP_1 embedding values and the boundary that you define
}
\description{
Create a dataframe with cell ID coordinates and cluster IDs from a Xenium Seurat object.
}
\examples{
distanceData_X2fov = object_FOV_to_coordinates(object = a, thisFOV = 'X2fov', threshold_y = 2694, angle_adjust = FALSE, theta_deg = 0, flip_x_coordinates = FALSE)
}
\author{
Margarita Kapustina
}
