% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Plot_geneExpressionvsMidline.R
\name{plotGeneExpressionVsMidline}
\alias{plotGeneExpressionVsMidline}
\title{Plots gene expression data away from the computed spatial midline (supports multiple FOVs)}
\usage{
plotGeneExpressionVsMidline(
  geneExpressionData,
  genes,
  binNumber = 7,
  binwidth = 40,
  save_plot = FALSE
)
}
\arguments{
\item{geneExpressionData}{List of dataframes generated with getExpressionvsMidline() across multiple FOVs}

\item{genes}{Character vector of gene names to their plot expression data}

\item{binNumber}{Number of bins to use for histogram (bars). Suggested binNumber = total # cells in pooled data/binwidth}

\item{binwidth}{Width of bins for gene expression averaging (lines) Suggested binwidth = 40 for 40micron bins}

\item{save_plot}{Option to save plot as .eps in working directory (TRUE, FALSE)}
}
\value{
Pooled dataframe containing Cell IDs, coordinates (X,Y), gene expression counts for specififed gene(s),
   and computed distance away from spatial midline of each cell, across all FOVs in data
}
\description{
Plot gene expression data for cells and their distance away from Spatial Midline across multiple FOVs.
}
\examples{
pooled_GeneExpression = plotGeneExpressionVsMidline(geneExpressionData = list(geneExpression_fov, geneExpression_fX1ov), genes = c("Epha1", "Cntnap4", 'Pcp4', 'Slc17a7'))
}
\author{
Margarita Kapustina
}
