% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetadataFrequency.R
\name{plotMetadataFrequency}
\alias{plotMetadataFrequency}
\title{Plot Cluster Frequencies and Proportions Stratified by Metadata}
\usage{
plotMetadataFrequency(
  obj,
  cluster_var,
  group_var,
  facet_var = NULL,
  colors = c("magenta", "grey")
)
}
\arguments{
\item{obj}{A Seurat object with metadata variables}

\item{cluster_var}{Character. Metadata variable containing cluster labels}

\item{group_var}{Character. Metadata variable used for grouping (fill variable)}

\item{facet_var}{Optional character. Metadata variable to facet plots by (Default is NULL)}

\item{colors}{Character vector of colors. Used if number of groups is <= length(colors)}
}
\value{
A list with two ggplot objects:
\item{raw}{Bar plot showing raw cell counts.}
\item{norm}{Bar plot showing proportions.}
}
\description{
Creates two bar plots showing (1) raw cell counts and (2) normalized proportions
of clusters across a metadata variable. Optionally facets by a third metadata variable.
}
\examples{
#generate plots
plots <- plotMetadataFrequency(obj = myObj, cluster_var = "seurat_clusters", group_var = "condition", facet_var = "sex")
#print plots
plots$raw + plots$norm

}
