% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP1_inSituMidline.R
\name{plotUMAP1_inSituMidline}
\alias{plotUMAP1_inSituMidline}
\title{Xenium 1-D UMAP Embeddings Plotted in Situ}
\usage{
plotUMAP1_inSituMidline(
  object,
  FOV = "X1fov",
  degs = 45,
  save_plot = FALSE,
  EmbeddingsPlotTitle = "UMAP_1 embedding for all\\n cells in subset",
  HistogramPlotTitle = "Distribution of UMAP_1 embedding\\n values all cells in subset",
  inSituPlotTitle = "UMAP_1 embedding in situ\\n for representative FOV"
)
}
\arguments{
\item{object}{a Xenium object}

\item{FOV}{FOV to plot UMAP_1 embedding values in}

\item{degs}{User-defined angle, defined in degrees, that intersects centrepoint in midline computation}

\item{save_plot}{Option to save plot as .eps in working directory (TRUE, FALSE)}

\item{EmbeddingsPlotTitle}{title for UMAP_1 Embeddings plot}

\item{HistogramPlotTitle}{title for UMAP_1 Embeddings Histogram plot}

\item{inSituPlotTitle}{title for UMAP_1 Embeddings in situ plot}
}
\value{
Dataframe containing Cell IDs, coordinates (X,Y) and computed UMAP_1 embeddings values for cells in specified FOV.
#Plots 1D UMAP, corresponding histogram of UMAP_1 embeddings values, and UMAP_1 embeddings values plotted in situ.
}
\description{
Plots: 1-dimensional UMAP, the corresponding histogram of UMAP_1 embedding values,
   the UMAP_1 embedding values in situ within specified FOV, and computes the midline intersecting cells within specified FOV.
   note: please compute UMAP_1 embeddings beforehand.
}
\examples{
AD_df = plotUMAP1_inSituMidline(object = AD, FOV = 'fov', degs = 65)
}
\author{
Margarita Kapustina
}
