% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP1_line_and_histogram.R
\name{plotUMAP1_vsMidline}
\alias{plotUMAP1_vsMidline}
\title{Plots UMAP_1 embeddings values for cells and their distance away from the computed spatial midline (supports multiple FOVs)}
\usage{
plotUMAP1_vsMidline(
  UMAP1MidlineData,
  binNumber = 7,
  binwidth = 40,
  save_plot = FALSE
)
}
\arguments{
\item{UMAP1MidlineData}{List of dataframes generated with getUMAP1_MidlineData() across multiple FOVs}

\item{binNumber}{Number of bins to use for histogram (bars). Suggested binNumber = total # cells in pooled data/binwidth}

\item{binwidth}{Width of bins for gene expression averaging (lines). Suggested binwidth = 40 for 40micron bins}

\item{save_plot}{Option to save plot as .eps in working directory (TRUE, FALSE)}
}
\value{
Pooled dataframe containing Cell IDs, coordinates (X,Y), UMAP_1 embedding values,
   and computed distance away from spatial midline of each cell, across all FOVs in data
}
\description{
Plot UMAP_1 embedding values for cells and their distance away from Spatial Midline across multiple FOVs.
}
\examples{
pooled_UMAP1_vsMidline = plotUMAP1_vsMidline(list(UMAP1_midline_data_fov, UMAP1_midline_data_X1fov))
}
\author{
Margarita Kapustina
}
