% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{XeniumBoxPlot}
\alias{XeniumBoxPlot}
\title{Xenium Gene Box Plots}
\usage{
XeniumBoxPlot(object, genes, ncol = 4, save_plot = FALSE)
}
\arguments{
\item{object}{Xenium object}

\item{genes}{Character vector of gene names to plot}

\item{ncol}{Specify number of columns in plot}

\item{save_plot}{Option to save plot as .eps in working directory (TRUE, FALSE)}
}
\value{
A grid of boxplots showing gene expression across all clusters
}
\description{
Create a boxplot denoting expression of gene(s) of interest across all clusters. SCT assay counts (sequencing depth corrected counts) are used.
}
\examples{
#create vector ie my_genes= c('C1ql2','Slc17a7', 'Gng13')
BoxPlots = XeniumBoxPlot(object = atn, genes = my_genes)
}
\author{
Margarita Kapustina
}
