% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotsRaw.R
\name{XeniumBoxPlotRaw}
\alias{XeniumBoxPlotRaw}
\title{Xenium Gene Box Plots - Raw Counts}
\usage{
XeniumBoxPlotRaw(genes, ncol = 4, object, save_plot = FALSE)
}
\arguments{
\item{genes}{Character vector of gene names to plot}

\item{ncol}{Specify number of columns in plot}

\item{object}{Xenium object}

\item{save_plot}{Option to save plot as .eps in working directory (TRUE, FALSE)}
}
\value{
A grid of boxplots showing raw gene expression across all clusters
}
\description{
Create a boxplot denoting expression of gene(s) of interest across all clusters. Xenium assay counts (raw counts) are used.
}
\examples{
#create vector ie my_genes= c('C1ql2','Slc17a7', 'Gng13')
BoxPlots = XeniumBoxPlotRaw(object = atn, genes = my_genes)
}
\author{
Margarita Kapustina
}
